Calculator app now supports float values rather than integers

+++ src/Calculator.java
@@ -7,8 +7,8 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface Calculator extends Remote {
-	int add(int num1, int num2) throws RemoteException;
-	int minus(int num1, int num2) throws RemoteException;
-	int multiply(int num1, int num2) throws RemoteException;
-	int divide(int num1, int num2) throws RemoteException;
+	float add(float num1, float num2) throws RemoteException;
+	float minus(float num1, float num2) throws RemoteException;
+	float multiply(float num1, float num2) throws RemoteException;
+	float divide(float num1, float num2) throws RemoteException;
 }
